fix: fix ExtractionUtils.extractionError，update lookup function
